+++ odbc-api/src/preallocated.rs
@@ -104,8 +104,8 @@ impl<'o> Preallocated<'o> {
     /// The resulting type is one level of indirection away from the raw pointer of the ODBC API. It
     /// no longer has any guarantees about bound buffers, but is still guaranteed to be a valid
     /// allocated statement handle. This serves together with
-    /// [`crate::handles::StatementImpl::into_sys`] or [`crate::handles::StatementImpl::as_sys`]
-    /// this serves as an escape hatch to access the functionality provided by `crate::sys` not yet
+    /// [`crate::handles::StatementImpl::into_sys`] or [`crate::handles::Statement::as_sys`] this
+    /// serves as an escape hatch to access the functionality provided by `crate::sys` not yet
     /// accessible through safe abstractions.
     pub fn into_statement(self) -> StatementImpl<'o> {
         self.statement