+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## Next
 
+* Add Support for sending long data from a binary slice.
 * Add variant `DataType::LongVarChar`.
 * Add variant `DataType::LongVarBinary`.
 * Add variant `Error::FailedReadingInput`.